+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Changed
 
-- `Innmind\Server\Status\ServerFactory::build()` now expects `Innmind\Server\Control\Server` as second argument and `Innmind\Immutable\Map<non-empty-string, string>` as third argument
+- `Innmind\Server\Status\ServerFactory::build()` now expects `Innmind\Server\Control\Server` as second argument and `Innmind\Server\Status\EnvironmentPath` as third argument
 
 ## 3.0.0 - 2022-01-23